advance Python self-host runtime and backend
Add and reorganize the self backend around target-specific parsing, dispatch, and emission paths for aarch64 Darwin and x86_64 Linux, with bootstrap gates, promotion gates, Docker coverage, and C testsuite runners.

Extend the Python frontend and codegen for native stdlib/runtime calls, CPython bridge marshalling, print/file/sys/os/path handling, class field access, dynamic isinstance, tuple iteration types, for-else/top-level control forms, and additional fallback ratchets.

Grow the Python runtime and substrate with PCC-emitted Python runtime modules plus C support for ints, strings, tuples, lists, dicts, sets, objects, exceptions, tracebacks, file/print/process/os/path handling, and runtime oracle coverage; stop checking in the generated libpy_runtime.a archive.

Rework CLI, packaging, and bootstrap scripts so pip/runtime builds can use the self backend path, and expand docs/plans around self-host gaps, GC semantics, Python data model coverage, native stdlib plans, fallback probes, and bootstrap status.

Add regression and corpus coverage for the new native paths, runtime oracle tests, bootstrap baselines, IR scaffold behavior, fallback baselines, Linux self-backend coverage, and project integration gates.

+++ pcc/py_runtime/src/py_dict.c
@@ -268,7 +268,7 @@ PyObject *py_dict_get_default(PyObject *dict, PyObject *key, PyObject *def) {
     return def;
 }
 
-int py_dict_contains(PyObject *dict, PyObject *key) {
+int64_t py_dict_contains(PyObject *dict, PyObject *key) {
     if (dict == NULL || key == NULL) return 0;
     PyDictObject *d = (PyDictObject *)dict;
     int64_t hash = py_obj_hash(key);
@@ -277,7 +277,7 @@ int py_dict_contains(PyObject *dict, PyObject *key) {
     return ix >= 0 ? 1 : 0;
 }
 
-int py_dict_del(PyObject *dict, PyObject *key) {
+int64_t py_dict_del(PyObject *dict, PyObject *key) {
     if (dict == NULL || key == NULL) return -1;
     PyDictObject *d = (PyDictObject *)dict;
     int64_t hash = py_obj_hash(key);
